% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-tables.R
\name{clean_etable}
\alias{clean_etable}
\title{Clean fixest etable Output}
\usage{
clean_etable(..., etable_args = list(), signif.col = FALSE)
}
\arguments{
\item{...}{One or more fixest model objects, or a list of fixest objects.}

\item{etable_args}{List of additional arguments passed to \code{\link[fixest:etable]{fixest::etable()}}.
Defaults include \code{tex = FALSE}, \code{se.below = TRUE}, \code{digits = 4}.}

\item{signif.col}{Logical. If TRUE, split significance stars into separate
columns adjacent to coefficients.}
}
\value{
A data.table with cleaned regression output.
}
\description{
Converts fixest etable output to a clean data.table format, optionally
splitting significance stars into separate columns.
}
\note{
Requires the \pkg{fixest} package.
}
\examples{
\dontrun{
library(fixest)
library(data.table)
data(mtcars)
m1 <- feols(mpg ~ hp, data = mtcars)
m2 <- feols(mpg ~ hp + wt, data = mtcars)
clean_etable(m1, m2)
}

}
