% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{empty_char_to_NA}
\alias{empty_char_to_NA}
\title{Convert Empty Strings to NA}
\usage{
empty_char_to_NA(dt)
}
\arguments{
\item{dt}{A data.table.}
}
\value{
The modified data.table (invisibly). The original object is
modified by reference.
}
\description{
Replaces empty strings ("") with \code{NA} in all character columns of a
data.table. Modifies the data.table by reference.
}
\examples{
library(data.table)
dt <- data.table(a = c("x", "", "z"), b = c("", "y", ""))
empty_char_to_NA(dt)
dt

}
