% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{most_frequent}
\alias{most_frequent}
\title{Find Most Frequent Value}
\usage{
most_frequent(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The most frequent value, or NA if the vector is empty.
}
\description{
Returns the most frequently occurring value in a vector.
}
\keyword{internal}
